+++ yazi-config/src/plugin/plugin.rs
@@ -1,12 +1,10 @@
-use std::sync::Arc;
-
 use anyhow::Result;
 use serde::{Deserialize, de};
 use yazi_codegen::DeserializeOver2;
 use yazi_shim::toml::DeserializeOverHook;
 
 use super::{Fetcher, Fetchers, Preloader, Preloaders, Previewer, Previewers, Spotter, Spotters};
-use crate::{mix, plugin::{FetcherArc, MAX_FETCHERS, MAX_PRELOADERS, PreloaderArc, PreviewerArc, SpotterArc}};
+use crate::{mix, plugin::{FetcherArc, PreloaderArc, PreviewerArc, SpotterArc}};
 
 #[derive(Default, Deserialize, DeserializeOver2)]
 pub struct Plugin {
@@ -37,32 +35,19 @@ pub struct Plugin {
 
 impl DeserializeOverHook for Plugin {
 	fn deserialize_over_hook(self) -> Result<Self, toml::de::Error> {
-		let mut fetchers: Vec<FetcherArc> =
+		let fetchers: Vec<FetcherArc> =
 			mix(self.prepend_fetchers, self.fetchers.unwrap_unchecked(), self.append_fetchers);
 		let spotters: Vec<SpotterArc> =
 			mix(self.prepend_spotters, self.spotters.unwrap_unchecked(), self.append_spotters);
-		let mut preloaders: Vec<PreloaderArc> =
+		let preloaders: Vec<PreloaderArc> =
 			mix(self.prepend_preloaders, self.preloaders.unwrap_unchecked(), self.append_preloaders);
 		let previewers: Vec<PreviewerArc> =
 			mix(self.prepend_previewers, self.previewers.unwrap_unchecked(), self.append_previewers);
 
-		if fetchers.len() > MAX_FETCHERS as usize {
-			Err(de::Error::custom(format!("Fetchers exceed the limit of {MAX_FETCHERS}")))?;
-		} else if preloaders.len() > MAX_PRELOADERS as usize {
-			Err(de::Error::custom(format!("Preloaders exceed the limit of {MAX_PRELOADERS}")))?;
-		}
-
-		for (i, p) in fetchers.iter_mut().enumerate() {
-			Arc::get_mut(p).ok_or_else(|| de::Error::custom("non-unique fetcher arc"))?.idx = i as u8;
-		}
-		for (i, p) in preloaders.iter_mut().enumerate() {
-			Arc::get_mut(p).ok_or_else(|| de::Error::custom("non-unique preloader arc"))?.idx = i as u8;
-		}
-
 		Ok(Self {
-			fetchers: fetchers.into(),
+			fetchers: fetchers.try_into().map_err(de::Error::custom)?,
 			spotters: spotters.into(),
-			preloaders: preloaders.into(),
+			preloaders: preloaders.try_into().map_err(de::Error::custom)?,
 			previewers: previewers.into(),
 			..Default::default()
 		})