+++ yazi-config/src/plugin/spotters.rs
@@ -1,13 +1,14 @@
 use std::{borrow::Cow, ops::Deref, sync::Arc};
 
 use arc_swap::ArcSwap;
-use mlua::{MetaMethod, UserData, UserDataMethods};
+use mlua::{ExternalError, ExternalResult, MetaMethod, UserData, UserDataMethods};
 use serde::Deserialize;
 use yazi_fs::file::File;
 use yazi_shared::id::Id;
-use yazi_shim::arc_swap::IntoPointee;
+use yazi_shim::{arc_swap::{ArcSwapExt, IntoPointee}, vec::{IndexAtError, VecExt}};
 
-use crate::plugin::{SpotterArc, SpotterMatcher};
+use super::Spotter;
+use crate::{mix, plugin::{SpotterArc, SpotterMatcher}};
 
 #[derive(Debug, Default, Deserialize)]
 pub struct Spotters(ArcSwap<Vec<SpotterArc>>);
@@ -41,6 +42,32 @@ impl Spotters {
 		}
 	}
 
+	pub fn insert(&self, index: isize, spotter: SpotterArc) -> Result<(), IndexAtError> {
+		self.0.try_rcu(|spotters| {
+			let i = spotters.index_at(index)?;
+			if i == spotters.len() {
+				Ok(mix(Vec::<Spotter>::new(), spotters.iter().cloned(), [spotter.clone()]))
+			} else {
+				let (before, after) = spotters.split_at(i);
+				Ok(mix(
+					Vec::<Spotter>::new(),
+					before.iter().cloned().chain([spotter.clone()]).chain(after.iter().cloned()),
+					Vec::<Spotter>::new(),
+				))
+			}
+		})?;
+
+		Ok(())
+	}
+
+	pub fn remove(&self, matcher: SpotterMatcher) {
+		self.0.rcu(|spotters| {
+			let mut next = Vec::clone(spotters);
+			next.retain(|spotter| !matcher.matches(spotter));
+			next
+		});
+	}
+
 	pub(crate) fn unwrap_unchecked(self) -> Vec<SpotterArc> {
 		Arc::try_unwrap(self.0.into_inner()).expect("unique spotters arc")
 	}
@@ -53,6 +80,22 @@ impl UserData for &'static Spotters {
 				Some(matcher) => matcher,
 				None => me.into(),
 			})
+		});
+
+		methods.add_method("insert", |_, &me, (index, spotter): (isize, SpotterArc)| {
+			let index = match index {
+				1.. => index - 1,
+				0 => return Err("index must be 1-based or negative".into_lua_err()),
+				_ => index,
+			};
+
+			me.insert(index, spotter.clone()).into_lua_err()?;
+			Ok(spotter)
+		});
+
+		methods.add_method("remove", |_, &me, matcher: SpotterMatcher| {
+			me.remove(matcher);
+			Ok(())
 		});
 
 		methods.add_meta_method(MetaMethod::Len, |_, me, ()| Ok(me.load().len()));